// src/api/models/userListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { UserListItemMainCarNumber } from './userListItemMainCarNumber';
import type { UserListItemMainCard } from './userListItemMainCard';

export interface UserListItem {
  /** ID */
  id: string;
  /** 이름 */
  name: string;
  /**
   * 이메일
   * @nullable
   */
  email?: string | null;
  /** 전화번호 */
  phoneNumber: string;
  /**
   * 주소
   * @nullable
   */
  address?: string | null;
  /** 추천인 코드 */
  referralCode: string;
  /**
   * 회원이 등록한 추천인 코드
   * @nullable
   */
  referrerCode?: string | null;
  /** 생성일 */
  createdAt: string;
  /**
   * 삭제일
   * @nullable
   */
  deletedAt?: string | null;
  /**
   * 정지일
   * @nullable
   */
  bannedAt?: string | null;
  /** 삭제 여부 */
  isDeleted: boolean;
  /** 정지 여부 */
  isBanned: boolean;
  /** 마케팅 수신 여부 */
  isMarketing: boolean;
  /** 대표차량번호 */
  mainCarNumber: UserListItemMainCarNumber;
  /** 대표카드 */
  mainCard: UserListItemMainCard;
}